r4更新
update to version r4

+++ Cpp/WorldGen.cpp
@@ -2,29 +2,29 @@
 #include "Def.h"
 //Perlin Noise 2D;
 namespace WorldGen{
-	float perm[256];
-	int   seed;
-	int   NoiseSmoothed = 1;
-	int   InterpolationMethod = INTERPOLATION_METHOD_POWER2;
-	float NoiseScaleX = 32;
-	float NoiseScaleZ = 32;
-	int   WaterLevel = 30;
+	double	perm[256];
+	int		seed;
+	int		NoiseSmoothed = 1;
+	int		InterpolationMethod = INTERPOLATION_METHOD_POWER2;
+	double	NoiseScaleX = 32;
+	double	NoiseScaleZ = 32;
+	int		WaterLevel = 30;
 
 	void perlinNoiseInit(int mapseed){
 		srand(mapseed);
 		int i;
 		for (i = 0; i != 256; i++){
-			perm[i] = (float)(rnd() * 256);
+			perm[i] = (double)(rnd() * 256);
 		}
 		seed = mapseed;
 	}
 
-	inline float Noise(int x, int y){
+	inline double Noise(int x, int y){
 		return perm[abs(x * 437 + y) % 256];
 	}
 
-	float SmoothedNoise(float x, float y){
-		float corners, sides, center;
+	double SmoothedNoise(double x, double y){
+		double corners, sides, center;
 		int intx = GoInt(x);
 		int inty = GoInt(y);
 		corners = (Noise(intx - 1, inty - 1) + Noise(intx + 1, inty - 1) + Noise(intx - 1, inty + 1) + Noise(intx + 1, inty + 1)) / 32;
@@ -33,13 +33,13 @@ namespace WorldGen{
 		return corners + sides + center;
 	}
 
-	float Interpolate(float a, float b, float x){
-		float temp, x2;
+	double Interpolate(double a, double b, double x){
+		double temp, x2;
 		switch (InterpolationMethod){
 		case INTERPOLATION_METHOD_LINEAR:
 			return a*(1 - x) + b*x;
 		case INTERPOLATION_METHOD_COSINE:
-			return float(a*(1 - (1 - cos(x*M_PI))*0.5) + b*(1 - cos(x*M_PI))*0.5);	
+			return double(a*(1 - (1 - cos(x*M_PI))*0.5) + b*(1 - cos(x*M_PI))*0.5);	
 		case INTERPOLATION_METHOD_POWER2:
 			temp = 1 - x;
 			temp *= temp;
@@ -54,18 +54,18 @@ namespace WorldGen{
 		}
 	}
 
-	float InterpolatedNoise(float x, float y){
+	double InterpolatedNoise(double x, double y){
 		int int_X, int_Y;
-		float fractional_X, fractional_Y, v1, v2, v3, v4, i1, i2;
+		double fractional_X, fractional_Y, v1, v2, v3, v4, i1, i2;
 		int_X = GoInt(x);
 		fractional_X = x - int_X;
 		int_Y = GoInt(y);
 		fractional_Y = y - int_Y;
 		if (NoiseSmoothed == 1){
-			v1 = SmoothedNoise((float)int_X, (float)int_Y);
-			v2 = SmoothedNoise((float)int_X + 1, (float)int_Y);
-			v3 = SmoothedNoise((float)int_X, (float)int_Y + 1);
-			v4 = SmoothedNoise((float)int_X + 1, (float)int_Y + 1);
+			v1 = SmoothedNoise((double)int_X, (double)int_Y);
+			v2 = SmoothedNoise((double)int_X + 1, (double)int_Y);
+			v3 = SmoothedNoise((double)int_X, (double)int_Y + 1);
+			v4 = SmoothedNoise((double)int_X + 1, (double)int_Y + 1);
 		}
 		else{
 			v1 = Noise(int_X, int_Y);
@@ -78,14 +78,14 @@ namespace WorldGen{
 		return Interpolate(i1, i2, fractional_Y);
 	}
 
-	float PerlinNoise2D(float x, float y){
-		float total, frequency, amplitude, a;
+	double PerlinNoise2D(double x, double y){
+		double total, frequency, amplitude, a;
 		int i, n;
 		total = 0;
 		n = 8;
 		for (i = 0; i <= n; i++){
-			frequency = (float)pow(2, i);
-			amplitude = (float)pow(0.5, i);
+			frequency = (double)pow(2, i);
+			amplitude = (double)pow(0.5, i);
 			a = InterpolatedNoise(x*frequency, y*frequency)*amplitude;
 			total += a;
 		}
@@ -93,12 +93,12 @@ namespace WorldGen{
 	}
 
 	int getHeight(int x, int y){
-		float NoiseHeight = PerlinNoise2D(x / NoiseScaleX, y / NoiseScaleZ) / 4.0f;
+		double NoiseHeight = PerlinNoise2D(x / NoiseScaleX, y / NoiseScaleZ) / 4.0f;
 		return int(NoiseHeight);
 	}
 
 	int getSandHeight(int x, int y){
-		float NoiseHeight = PerlinNoise2D(x / NoiseScaleX * 3 + 1222, y / NoiseScaleZ * 3 + 1222) / 4.0f - 2;
+		double NoiseHeight = PerlinNoise2D(x / NoiseScaleX * 3 + 1222, y / NoiseScaleZ * 3 + 1222) / 4.0f - 2;
 		return int(NoiseHeight);
 	}
 }
